# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "This is the API documentation for Safar API",
        "contact": {
            "email": "zahra77babaee@gmail.com.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/notifications": {
            "get": {
                "tags": [
                    "اعلانات"
                ],
                "summary": "دریافت اعلانات کاربر",
                "description": "این متد تمام اعلانات مربوط به کاربر را نمایش می‌دهد. با استفاده از پارامتر `is_read` می‌توان اعلانات را بر اساس خوانده شده یا نشده بودن فیلتر کرد.",
                "operationId": "3bcce2782747db8b0c7a14103ce2b7f9",
                "parameters": [
                    {
                        "name": "is_read",
                        "in": "query",
                        "description": "فیلتر وضعیت خوانده شده (0=خوانده نشده، 1=خوانده شده)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "لیست اعلانات",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "خوش آمدید"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "به سایت ما خوش آمدید"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "wellcome"
                                                    },
                                                    "is_read": {
                                                        "type": "boolean",
                                                        "example": false
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1//notifications/{id}/read": {
            "patch": {
                "tags": [
                    "اعلانات"
                ],
                "summary": "علامت‌گذاری اعلان به عنوان خوانده شده",
                "description": "این متد یک اعلان خاص را برای کاربر به عنوان خوانده شده علامت‌گذاری می‌کند.",
                "operationId": "6bc111f1ba0cd71e26d2962627884e9a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "شناسه اعلان",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "اعلان به عنوان خوانده شده علامت‌گذاری شد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Notification marked as read"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "اعلان یافت نشد"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
...
}